# perf: replace the ipv6 map by an array of struct