While starting `ipfs` this package used to take ~20ms while initing:
```
init github.com/libp2p/go-libp2p-asn-util @21 ms, 20 ms clock, 2623128 bytes, 892 allocs
```
That is because the map literal is actually creating a new map and adding entries one by one, which for such big map takes a while.
This PR moves this to a separate goroutine and use a `sync.Once` to synchronise accesses. Almost completely removing it:
```
init github.com/libp2p/go-libp2p-asn-util @8.7 ms, 0.002 ms clock, 128 bytes, 3 allocs
```