@Jorropo 2022-04-12
@willscott let's not merge this. I have red the code and this change is overkill.

Never is this map actually used as a map, the only consumer just iterate over it.

So we can use an array of structs containing `k` and `v` instead.