@Jorropo 2022-04-12
@willscott Update it's now slower `0.005ms` instead of `0.002ms` however it's `0.005ms` of CPU time instead of `0.002ms` of wall time and `20ms` of CPU time.